+++ src/StackExchange.Redis/Enums/RedisCommand.cs
@@ -85,6 +85,7 @@
         LLEN,
         LMOVE,
         LPOP,
+        LPOS,
         LPUSH,
         LPUSHX,
         LRANGE,